requirements.txt
fastapi
uvicorn
python-multipart
tensorflow
scikit-learn
pandas
librosa
soundfile
praat-parselmouth
webrtcvad-wheels
numpy